extract icon test
