+++ src/l-icon.js
@@ -1,5 +1,4 @@
-// @vitest-environment happy-dom
-import * as L from "leaflet";
+import { icon } from "leaflet";
 import {
   bool,
   chain,
@@ -40,13 +39,19 @@ class LIcon extends HTMLElement {
       "popup-anchor",
     ];
     points.forEach((key) => {
-      obj[kebabToCamel(key)] = chain(optional(htmlAttribute(key)), nullable(json()));
+      obj[kebabToCamel(key)] = chain(
+        optional(htmlAttribute(key)),
+        nullable(json())
+      );
     });
-    obj["crossOrigin"] = chain(optional(htmlAttribute("cross-origin")), nullable(bool()));
+    obj["crossOrigin"] = chain(
+      optional(htmlAttribute("cross-origin")),
+      nullable(bool())
+    );
     const schema = partial(obj);
     const options = parse(schema, this);
 
-    this.icon = L.icon(options);
+    this.icon = icon(options);
 
     const event = new CustomEvent("icon:add", {
       cancelable: true,
@@ -57,70 +62,6 @@ class LIcon extends HTMLElement {
     });
     this.dispatchEvent(event);
   }
-}
-
-if (import.meta.vitest) {
-  const { it, expect, beforeAll } = import.meta.vitest;
-
-  beforeAll(() => {
-    customElements.define("l-icon", LIcon);
-  });
-
-  it("default", () => {
-    const el = document.createElement("l-icon");
-    document.body.appendChild(el);
-
-    let actual = el.icon;
-    let expected = L.icon();
-    expect(actual).toEqual(expected);
-  });
-
-  it("emits icon:add event", async () => {
-    const el = document.createElement("l-icon");
-    let promise = new Promise((resolve) => {
-      el.addEventListener("icon:add", (ev) => {
-        resolve(ev.detail.icon);
-      });
-    });
-    document.body.appendChild(el);
-    let actual = await promise;
-    let expected = L.icon();
-    expect(actual).toEqual(expected);
-  });
-
-  it("options", () => {
-    const el = document.createElement("l-icon");
-    el.setAttribute("icon-url", "url.png");
-    el.setAttribute("icon-retina-url", "retina.png");
-    el.setAttribute("icon-size", "[0, 0]");
-    el.setAttribute("icon-anchor", "[0, 0]");
-    el.setAttribute("popup-anchor", "[0, 0]");
-    el.setAttribute("tooltip-anchor", "[0, 0]");
-    el.setAttribute("shadow-url", "urlShadow.png");
-    el.setAttribute("shadow-retina-url", "retinaShadow.png");
-    el.setAttribute("shadow-size", "[0, 0]");
-    el.setAttribute("shadow-anchor", "[0, 0]");
-    el.setAttribute("class-name", "foo");
-    el.setAttribute("cross-origin", "true");
-    document.body.appendChild(el);
-
-    let actual = el.icon;
-    let expected = L.icon({
-      iconUrl: "url.png",
-      iconRetinaUrl: "retina.png",
-      iconSize: [0, 0],
-      iconAnchor: [0, 0],
-      popupAnchor: [0, 0],
-      tooltipAnchor: [0, 0],
-      shadowUrl: "urlShadow.png",
-      shadowRetinaUrl: "retinaShadow.png",
-      shadowSize: [0, 0],
-      shadowAnchor: [0, 0],
-      className: "foo",
-      crossOrigin: true,
-    });
-    expect(actual).toEqual(expected);
-  });
 }
 
 export default LIcon;
